requirements/common.txt
# https://pypi.python.org/pypi/Django
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
django-filter==2.4.0
